third_party/WebKit/LayoutTests/fast/dom/document-set-title-no-reuse.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script type="text/javascript">
description("Test that setting document.title does not reuse &lt;title&gt;'s textnode child");

onload = function()
{
    document.title = "aaa";
    oldnode = document.getElementsByTagName("title")[0].firstChild;
    document.title = "bbb";
    shouldBeTrue('oldnode.textContent != "bbb"');
    shouldBeTrue('oldnode != document.getElementsByTagName("title")[0].firstChild');
}
</script>
