third_party/WebKit/LayoutTests/fast/dom/document-set-title-no-reuse-expected.txt

Test that setting document.title does not reuse <title>'s textnode child

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS oldnode.textContent != "bbb" is true
PASS oldnode != document.getElementsByTagName("title")[0].firstChild is true
